feat: type intersection option for `allOf` and `anyOf` schema
A new option (`-i`/`--intersection`) enables output of [TypeScript intersection
types](https://www.typescriptlang.org/docs/handbook/advanced-types.html) for
`allOf` and `anyOf` schema. Specifically, `allOf` schema are rendered as types
defined as, e.g. `A & B` and `anyOf` schema as `(A | B)` (the parentheses in
the latter to assert precedence).

As a byproduct of these changes, the rendering process now outputs "/* ..*/"
comments for the last enum value in the list, consistent with prior items in
the enum, whereas it had previously emitted the last comment in "//" style, and
so this change also includes a corresponding tweak to the test fixture in `advanced_schema.d.ts`.