Add support for client-specified namespace
When running dtsgenerator on an OpenAPI 2.0 (f.k.a Swagger) input
file, the schemas are nested under an object named "definitions".
In OpenAPI 3.0, schemas are under "components" -> "schemas".
This option allows the caller to override this and specify an
explicit namespace (and thus have the same output, no matter which
version of OpenAPI is used), rather than these artificial names
which are unrelated to the content.

Via the CLI, use `--namespace <namespace>` or `-n <namespace>`.
If the `<namespace>` is the special name `~none~`, then no namespace
is put around the types.