refactor: use `export` when given empty namespace
Our use case is that we're generating a package with from the OpenAPI
and given `--namespace ''` the generated interfaces are declared and not
exported. We could specify a namespace instead and the namespace would
be declared and the interfaces exported, but then we would need to use
the `namespace.InterfaceName` syntax or use cumbersome imports like
`import InterfaceName = namespace.InterfaceName` or type definitions
like `type InterfaceName = namespace.InterfaceName` throught the code
base.

Right now, we can either add those imports/type definitions or we can
run `sed` on the generated TypeScript file to change `declare` to
`export` as a workaround.

So, when using with empty namespace, i.e. specifying `--namespace ''` on
the CLI the generated interfaces should be exported as well. This adds a
condition for such a case by simply checking if given namespace is
empty.